Update patch set 2
Patch Set 2:

(2 comments)

What tests did you do to ensure that the generated stubs are identical to the previously generated stubs? Relying on the build to detect is not sufficient as it will only not detect when things are missing not if extra methods/fields/classes are output.

Patch-set: 2
Reviewer: Paul Duffin <1039338@85c56323-6fa9-3386-8a01-6480fb634889>