org.apache.http.legacy is built with java_sdk_library
The runtime library, stubs library and droiddoc are all abstracted into
the java_sdk_library module definition.

note: as a small-side effect, X509TrustManagerExtensions from
frameworks/base is included as the API signature of the legacy library.
This is because java_sdk_library includes all java classes from the
frameworks/base that are in the packages listed in the property
api_packages.

However, this isn't a breaking change because the added class signature
is already part of Android API (and we already have duplicated API
signatures across org.apache.http.legacy and Android API: SslCertficate,
SslError, etc)

Bug: 77577799

Test: m -j

Change-Id: I30ce741d7dfed1f93d3b19c750dcfbe073cceb69